Move tags cell renderer into a separate file for re-use
